@@ -3,7 +3,7 @@
 	"description": "IntelliSense for skript",
 	"author": "John Heikens",
 	"license": "GNU",
-	"version": "1.7.0",
+	"version": "1.7.2",
 	"icon": "icon.png",
 	"repository": {
 		"type": "git",
@@ -23,7 +23,6 @@
 		"vscode": "^1.63.0"
 	},
 	"activationEvents": [
-		"onLanguage:skript",
 		"onLanguage:Sk-VSC",
 		"onLanguage:sk",
 		"onLanguage:skriptlang"
@@ -38,10 +37,12 @@
 				],
 				"aliases": [
 					"Skript",
-					"sk"
+					"sk",
+					"Sk-VSC",
+					"skriptlang"
 				],
 				"filenames": [],
-				"firstLine": "#sk",
+				"firstLine": "^#\\s*Skript",
 				"configuration": "./language-configuration.json",
 				"icon": {
 					"light": "./skript icon.png",
@@ -81,8 +82,8 @@
 	},
 	"scripts": {
 		"vscode:prepublish": "npm run compile",
-		"compile": "tsc -b",
-		"watch": "tsc -b -w",
+		"compile": "npx tsc -b",
+		"watch": "npx tsc -b -w",
 		"lint": "eslint ./client/src ./server/src --ext .ts,.tsx",
 		"postinstall": "cd client && npm install && cd ../server && npm install && cd ..",
 		"test": "sh ./scripts/e2e.sh"
@@ -94,7 +95,7 @@
 		"@typescript-eslint/parser": "^5.42.0",
 		"eslint": "^8.26.0",
 		"mocha": "^9.2.1",
-		"typescript": "^4.8.4"
+		"typescript": "^4.9.5"
 	},
 	"dependencies": {
 		"vscode-uri": "^3.0.7"
