@@ -1,4 +1,4 @@
-import { PatternData } from './PatternData';
+import { PatternData } from './Data/PatternData';
 
 
 export type PatternResultProcessor = (patternFound: PatternData) => boolean;
