@@ -1,5 +1,5 @@
 import { NestHierarchy } from '../Nesting/NestHierarchy';
-import { PatternData } from "../Pattern/PatternData";
+import { PatternData } from "../Pattern/Data/PatternData";
 export class SkriptPatternMatchHierarchy extends NestHierarchy<SkriptPatternMatchHierarchy>{
 	matchedPattern?: PatternData;
 	constructor(start?: number, end?: number, matchedPattern?: PatternData) {
